services/inference/requirements.txt
fastapi
uvicorn
pandas
numpy
tensorflow
python-jose[cryptography]
python-multipart
mlflow>=2.0.0